Fix cheats for libretro
The cheats were not integrated properly, as a result, only the last
cheat was applied. This is because gamebatte expects all cheats to be
set in one go and separated by a semi-colon. But the libretro api wants
to apply cheats one at a time. Therefore every time a cheat is applied
it disables the previous one.

Also clearCheats was not properly implemented at all, it did not take
into account gameshark codes and it was simply not working for gamegenie,
therefore it was impossible to disable a cheat without reloading the core.