fix sound on platforms where 'unsigned long' is only 32 bits wide
Commit 1d5f921d1846a2939b83334157b30dd7a34e80c8 introduced a bit shift
that is too wide on 32 bit platforms. This results in the volume of
all channels being set to zero in the internal mixer.

Problem observed on a Raspberry Pi Zero W, where the compiler also
complains:

libgambatte/src/sound.cpp:106:63:
  warning: right shift count >= width of type [-Wshift-count-overflow]
  ch1_.update(buf, (soChVol_[0] * soVol_ * 0x1999999A) >> 32, cycles);
                                                          ^~

Fix this problem by expanding the constant 0x1999999A to unsigned long long
via the 'ULL' suffix. This results in all values in the expression being
promoted to 'unsigned long long', which is 64 bit wide everywhere.