2024 New Year's Patch
Happy New Year to the gambatte community!

2024 marks 35 years of the Nintendo Game Boy, a handheld that revolutionized the portable gaming market. And to commemorate, as of this patch, almost all Game Boy palettes that you've been using for your portable adventures, especially on the Miyoo Mini (or the Miyoo Mini Plus) have received a brighter makeover via color-hex!

To give context: I used to visit a Japanese website called SYNCER to use its Color Explorer tool which made my Game Boy palettes possible. Unfortunately, the Color Explorer tool no longer exists. So to cope with this loss, I have decided to move over to color-hex. So with this update, almost all of my Game Boy palettes are going to become more brighter.

In addition to this, 52 already-existing palettes are going to be overhauled entirely! Most of the 52 are now multi-colored, while some of them are now two-toned! Which palettes have changed altogether? You'll just have to find out!